Annotation update

+++ annotations/nocedit.exe/pseudocode/include/Nocturne/FunctionDefinition/Game/CSoundDevice_set3DListenerPos.h
@@ -3,5 +3,5 @@
 // Individual function definition header for: CSoundDevice_set3DListenerPos
 
 // Function Definition: CSoundDevice_set3DListenerPos
-typedef void (*CSoundDevice_set3DListenerPos)(CSoundDevice* this_ptr, float x, float y, float z);
+typedef void (*CSoundDevice_set3DListenerPos)(CSoundDevice* this_ptr, double x, double y, double z);
